Abstract File System should be fully working now. RefString.Split now has a convenient version called RefStringSplitQuery to work with foreach(...).
