+++ README.md
@@ -34,6 +34,7 @@ Also, the library is aimed on providing functions in pure C# and compatible with
 |GeneralPurposeScanner		| A general purpose scanner. |
 |TreeModel					| A general purpose TreeModel with Xml support. |
 |RefString					| A struct to help reduce GC when dealing with string. |
+|Abstract File System		| An extendable mapping file system. |
 
 ## License
