+++ Samples/AbstractFileSystemSamples/Program.cs
@@ -1,4 +1,5 @@
 ﻿using LibCLCC.NET.AbstractFileSystem;
+using LibCLCC.NET.Data;
 
 namespace AbstractFileSystemSamples
 {
@@ -16,36 +17,53 @@ namespace AbstractFileSystemSamples
             root.MapDirectory("/home/", homeProvider);
             if (root.GetFile("/home/.bashrc", out var file))
             {
-                WriteExampleFile(file);
+                WriteExampleFile(file, "PS1=\"\\w\\h:\"\n");
             }
             else Console.WriteLine("Failed Successfully.");
 
-            if (root.GetDirectory("/home/creeperlv/", out var dir))
             {
-                if (dir.TryCreate())
+
+                if (root.GetDirectory("/home/", out var dir))
                 {
-                    Console.WriteLine("Created home directory.");
+                    var result = dir.EnumerateDirectories();
+                    while (result.MoveNext())
+                    {
+                        Console.WriteLine(result.Current.FilePath);
+                    }
                 }
-                else
+            }
+            {
+
+                if (root.GetDirectory("/home/creeperlv/", out var dir))
                 {
-                    Console.WriteLine("Fail to create directory.");
+                    if (dir.TryCreate())
+                    {
+                        Console.WriteLine("Created home directory.");
+                    }
+                    else
+                    {
+                        Console.WriteLine("Fail to create directory.");
+                    }
                 }
             }
             if (root.GetFile("/home/creeperlv/.bashrc", out var file1))
             {
-                WriteExampleFile(file1);
+                WriteExampleFile(file1, "PS1=\"\\w\\h:\"\nSecond Line.\nThird Line");
+                file1.TryDelete();
+                WriteExampleFile(file1, "Second Write.\n");
+
             }
             else Console.WriteLine("Fail.");
         }
 
-        private static void WriteExampleFile(FileDescriptor file)
+        private static void WriteExampleFile(FileDescriptor file, string value)
         {
             if (file?.TryOpen(FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.ReadWrite, out var stream) ?? false)
             {
                 if (stream is not null)
                 {
-                    using StreamWriter SW = new StreamWriter(stream);
-                    SW.Write("""PS1="\\w\\h:"\n""");
+                    using StreamWriter SW = new(stream);
+                    SW.Write(value);
                     SW.Flush();
                 }
             }
